made the section components separated from the templates so they can be reused
