--- src/components/editor/add-page-button/AddPage.module.css
@@ -1,28 +1,0 @@
-.container {
-    width: 70%;
-    min-width: 309px;
-    max-width: 850px;
-    margin-inline: auto;
-    margin-block-start: var(--gap-medium);
-}
-
-.button {
-    width: 100%;
-    height: 2rem;
-    background-color: transparent;
-    border: 2px dashed #e1e1e1;
-    border-radius: var(--border-radius-xsmall);
-    color: #717171;
-    font-size: var(--font-size-small);
-    cursor: pointer;
-}
-
-.button:hover {
-    background-color: white;
-}
-
-@media screen and (max-width: 768px) {
-    .container {
-        width: 90%;
-    }
-}
