+++ src/components/editor/document-container/DocumentContainer.tsx
@@ -6,14 +6,13 @@ import { useAppContext } from "@/app/providers";
 import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
 import LoadingScreen from "@/components/loading-screen/LoadingScreen";
-import { updateDocumentArray } from "@/lib/document";
 
 type DocumentContainerProps = {
-    resume: any;
+    document: any;
 };
 
-const DocumentContainer = ({ resume }: DocumentContainerProps) => {
-    const id = resume.id;
+const DocumentContainer = ({ document }: DocumentContainerProps) => {
+    const id = document.id;
     const { documentArray, setDocumentArray } = useAppContext();
     const [currentDocument, setCurrentDocument] = useState<any>(null);
 
@@ -24,18 +23,18 @@ const DocumentContainer = ({ resume }: DocumentContainerProps) => {
     }, [documentArray]);
 
     useEffect(() => {
-        // add the resume to the document array if it is not already there
+        // add the document to the document array if it is not already there
         const tempArray = [...documentArray];
         if (!tempArray.some((document: any) => document.id === id)) {
             tempArray.push({
                 id,
                 currentPage: 1,
-                information: resume.information,
+                information: document.information,
             });
             setCurrentDocument({
                 id,
                 currentPage: 1,
-                information: resume.information,
+                information: document.information,
             });
             setDocumentArray(tempArray);
         } else {
@@ -46,35 +45,23 @@ const DocumentContainer = ({ resume }: DocumentContainerProps) => {
         }
     }, []);
 
-    const updateDocument = (updatedDocument: any) => {
-        const newDocumentArray = updateDocumentArray(
-            updatedDocument,
-            documentArray
-        );
-        setDocumentArray(newDocumentArray);
-    };
-
     const documentPages =
         currentDocument &&
         Array.from({ length: currentDocument.information.numPages }).map(
             (_, index) => (
-                <Basic
-                    key={`page${index}`}
-                    resumeId={id}
-                    updateDocument={updateDocument}
-                    index={index}
-                />
+                <Basic key={`page${index}`} resumeId={id} index={index} />
             )
         );
 
     return (
         <section className={styles.documentContainer}>
             <section className={styles.document}>
-                <DndProvider backend={HTML5Backend}>
-                    {!currentDocument && <LoadingScreen />}
-                    {currentDocument &&
-                        documentPages[currentDocument.currentPage - 1]}
-                </DndProvider>
+                {!currentDocument && <LoadingScreen />}
+                {currentDocument && (
+                    <DndProvider backend={HTML5Backend}>
+                        {documentPages[currentDocument.currentPage - 1]}
+                    </DndProvider>
+                )}
             </section>
         </section>
     );
